backend/requirements.txt
fastapi
uvicorn
telethon
aiomysql
python-dotenv
pydantic
